Extract mergeTest
